refactor the ErrorInputFrame
