+++ src/view/FileChooser.java
@@ -2,12 +2,10 @@ package view;
 
 import java.io.*;
 
-import java.awt.*;
-import java.awt.event.*;
 import javax.swing.*;
 import javax.swing.filechooser.FileNameExtensionFilter;
 
-import view.components.ColoredButton;
+import view.components.*;
 
 public class FileChooser extends JFileChooser {
     File choosenFile = null;
@@ -39,61 +37,6 @@ public class FileChooser extends JFileChooser {
             // stop.
             this.choosenFile = null;
         }
-    }
-
-    public class ErrorInputFrame extends JFrame {
-        private ColoredButton label = new ColoredButton("Wrong input file");
-        private JPanel buttonPanel = new JPanel();
-        private ColoredButton againButton = new ColoredButton("Choose again");
-        private ColoredButton exitButton = new ColoredButton("Exit");
-
-        public ErrorInputFrame() {
-            againButton.addActionListener(new ButtonListener());
-            exitButton.addActionListener(new ButtonListener());
-            this.setLayout(new GridBagLayout());
-            buttonPanel.setLayout(new GridBagLayout());
-            GridBagConstraints gbc = new GridBagConstraints();
-
-            gbc.gridx = 0;
-            gbc.gridy = 0;
-            this.add(label, gbc);
-            gbc.fill = GridBagConstraints.HORIZONTAL;
-
-            gbc.gridx = 0;
-            gbc.gridy = 0;
-            buttonPanel.add(againButton, gbc);
-
-            gbc.gridx = 1;
-            gbc.gridy = 0;
-            buttonPanel.add(exitButton, gbc);
-
-            gbc.gridx = 0;
-            gbc.gridy = 1;
-            this.add(buttonPanel, gbc);
-            gbc.fill = GridBagConstraints.HORIZONTAL;
-
-            this.setDefaultCloseOperation(JFrame.DISPOSE_ON_CLOSE);
-            this.setLocationRelativeTo(null);
-            this.pack();
-            this.setVisible(true);
-        }
-
-        private class ButtonListener implements ActionListener {
-            public void actionPerformed(ActionEvent e) {
-                if (e.getSource() == exitButton) {
-                    DisposeErrorInputFrame();
-                }
-                if (e.getSource() == againButton) {
-                    DisposeErrorInputFrame();
-                    new FileChooser();
-                }
-            }
-        }
-
-        private void DisposeErrorInputFrame() {
-            this.dispose();
-        }
-
     }
 
     // The getFile() method return the path of the choosen file, which will be used
